Update to add a line of comment to the file.
This python file is dependent on another file to function. Without the previous file, this python application would not actually function. Likewise, if the dictionary file(will be in this repository as well) were corrupted or had errors in it, the python file would be more susceptible to errors and/or crashing.